Add support for WPA CLI zephyr
In Linux WPA CLI is a standalone application that uses sockets to
communicate with ctrl_iface (uses Eloop), in Zephyr we have replaced
sockets with direct function calls, but this introduces concurrency as
we don't go through Eloop which causes instability and crashes esp. in
the deinitialization path.

Implement Zephyr's version of WPA CLI and control interface using
socketpair UNIX style sockets, here WPA CLI can be invoked as a standard
function rather than an executable. To avoid code duplication move all
commands processing to a separate file that can be used by wpa_cli
application and also wpa_cli_zephyr library (all other changes to
wpa_cli are restored).
