+++ src/common/wpa_ctrl.h
@@ -12,6 +12,8 @@
 #ifdef  __cplusplus
 extern "C" {
 #endif
+extern struct wpa_ctrl *ctrl_conn;
+extern char *ifname_prefix;
 
 /* wpa_supplicant control interface - fixed message prefixes */
 
@@ -476,6 +478,17 @@ enum wpa_vendor_elem_frame {
 
 /* wpa_supplicant/hostapd control interface access */
 
+#ifdef CONFIG_CTRL_IFACE_ZEPHYR
+/**
+ * wpa_ctrl_open - Open a control interface to wpa_supplicant/hostapd
+ * @sock: one of the pair of UNIX domain socket (socketpair)
+ * Returns: Pointer to abstract control interface data or %NULL on failure
+ *
+ * This function is used to open a control interface to wpa_supplicant/hostapd
+ * using a connected socketpair.
+ */
+struct wpa_ctrl * wpa_ctrl_open(const int sock);
+#else
 /**
  * wpa_ctrl_open - Open a control interface to wpa_supplicant/hostapd
  * @ctrl_path: Path for UNIX domain sockets; ignored if UDP sockets are used.
@@ -487,6 +500,7 @@ enum wpa_vendor_elem_frame {
  * interface need to use matching path configuration.
  */
 struct wpa_ctrl * wpa_ctrl_open(const char *ctrl_path);
+#endif
 
 /**
  * wpa_ctrl_open2 - Open a control interface to wpa_supplicant/hostapd
@@ -634,6 +648,8 @@ void wpa_ctrl_cleanup(void);
 char * wpa_ctrl_get_remote_ifname(struct wpa_ctrl *ctrl);
 #endif /* CONFIG_CTRL_IFACE_UDP */
 
+int wpa_ctrl_command(struct wpa_ctrl *ctrl, const char *cmd);
+int wpa_request(struct wpa_ctrl *ctrl, int argc, char *argv[]);
 
 #ifdef  __cplusplus
 }
