+++ zephyr/CMakeLists.txt
@@ -35,7 +35,7 @@ zephyr_library_compile_definitions(
 	CONFIG_WPA_S_ZEPHYR_L2_WIFI_MGMT
 	#CONFIG_NO_STDOUT_DEBUG
 	CONFIG_CTRL_IFACE
-	CONFIG_CTRL_IFACE_UDP
+	CONFIG_CTRL_IFACE_ZEPHYR
 	CONFIG_NO_RANDOM_POOL
 	CONFIG_MBO
 	CONFIG_WNM
@@ -69,7 +69,7 @@ zephyr_library_sources(
 	${COMMON_SRC_BASE}/utils/eloop.c
 	${COMMON_SRC_BASE}/utils/os_zephyr.c
 	${COMMON_SRC_BASE}/utils/radiotap.c
-	#${COMMON_SRC_BASE}/utils/edit_simple.c
+	${COMMON_SRC_BASE}/utils/edit_simple.c
 	${WPA_SUPPLICANT_BASE}/config.c
 	${WPA_SUPPLICANT_BASE}/notify.c
 	${WPA_SUPPLICANT_BASE}/bss.c
@@ -86,10 +86,11 @@ zephyr_library_sources(
 	${WPA_SUPPLICANT_BASE}/scan.c
 	${WPA_SUPPLICANT_BASE}/robust_av.c
 	${WPA_SUPPLICANT_BASE}/ctrl_iface.c
-	${WPA_SUPPLICANT_BASE}/ctrl_iface_udp.c
-	${WPA_SUPPLICANT_BASE}/wpa_cli.c
 	${WPA_SUPPLICANT_BASE}/mbo.c
 	${WPA_SUPPLICANT_BASE}/wnm_sta.c
+	${WPA_SUPPLICANT_BASE}/wpa_cli_cmds.c
+	${WPA_SUPPLICANT_BASE}/ctrl_iface_zephyr.c
+	${WPA_SUPPLICANT_BASE}/wpa_cli_zephyr.c
 	# Zephyr main
 	src/supp_main.c
 	src/utils/wpa_debug.c
